requirements.txt
albumentations==1.3.1
audioread==3.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.4.2
decorator==5.1.1
dlib==19.24.8
einops==0.6.1
idna==3.10
imageio==2.31.2
importlib-metadata==6.7.0
joblib==1.3.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.39.1
msgpack==1.0.5
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pip==22.3.1
platformdirs==4.0.0
pooch==1.8.2
pycparser==2.21
python-dateutil==2.9.0.post0
python-speech-features==0.6
pytz==2025.2
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools==65.6.3
six==1.17.0
soundfile==0.13.1
soxr==0.3.7
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==2.0.7
wheel==0.38.4
zipp==3.15.0
